Maintenance: mainly javadocs, but also some code improvements. This is not exhaustive. Done with the help of Intellij IDEA Code Inspection tools

+++ src/ares/application/shared/models/forces/IdentifiedUnitModel.java
@@ -31,7 +31,7 @@ public class IdentifiedUnitModel extends DetectedUnitModel {
     }
 
     public int getHealth() {
-        return (int) (unit.getEfficacy() - 1 / 20);
+        return unit.getEfficacy() - 1 / 20;
     }
 
     public Echelon getEchelon() {
